backend/requirements.txt
fastapi==0.88.0
psycopg2==2.9.5
pymongo==4.3.3
python-dotenv==0.21.0
uvicorn==0.20.0